requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
mysql-connector-python>=9.0.0
APScheduler>=3.10.0
pytz>=2024.1




